+++ hw4/BAH_Predictor.hpp
@@ -11,7 +11,7 @@ class BAH_Predictor : public Predictor
 {
     private:
     int table[TABLE_SIZE] = {};
-    int col[TABLE_SIZE] = {}; //keep track of the collision, store the upper bits
+    unsigned long long col[TABLE_SIZE] = {}; //keep track of the collision, store the upper bits
     bool colided;
 
 
@@ -19,11 +19,11 @@ class BAH_Predictor : public Predictor
     /* Constructor */
     BAH_Predictor();
 
-    bool get_result(int branch, int target, bool f);
+    bool get_result(unsigned long long branch, unsigned long long target, bool f);
 
-    bool predict(int branch, int target);
+    bool predict(unsigned long long branch, unsigned long long target);
 
-    bool collided(int branch); 
+    bool collided(unsigned long long branch); 
 
     /* Destructor */
     ~BAH_Predictor() = default;
