+++ hw4/SAT_Predictor.hpp
@@ -9,18 +9,18 @@ class SAT_Predictor : public Predictor
 {
   private:
     int table[TABLE_SIZE] = {};
-    int col[TABLE_SIZE] = {}; //keep track of the collision, store the upper bits
+    unsigned long long col[TABLE_SIZE] = {}; //keep track of the collision, store the upper bits
     bool colided;
 
   public:
     /* Constructor */
     SAT_Predictor();
 
-    bool get_result(int branch, int target, bool f);
+    bool get_result(unsigned long long branch, unsigned long long target, bool f);
 
-    bool predict(int branch, int target);
+    bool predict(unsigned long long branch, unsigned long long target);
 
-    bool collided(int branch);
+    bool collided(unsigned long long branch);
     /* Destructor */
     ~SAT_Predictor() = default;
 };
